Consolidate version checking in one place
We could be defensive and raise an error if we detect e.g. Live 13. On
the other hand, we could just try to parse it anyway and see if it
works. I suspect this code might not change much, so there a chance Live
13 could be released and no action is required on our part. Of course it
would be nice to add tests when that happens, but I'd rather not
be so strict and guarantee that we break when L13 comes out.